DragNDrop
Added more utility to the drag and drop function of inventories. Can now drag half stack or drop one from stack.

+++ src/warped/application/gui/GUIInventory.java
@@ -4,12 +4,15 @@ package warped.application.gui;
 import java.awt.Color;
 import java.awt.Font;
 import java.awt.Graphics;
+import java.awt.event.KeyEvent;
+import java.awt.event.MouseEvent;
+import java.util.HashMap;
 
 import warped.application.entities.item.ItemBindable;
 import warped.application.entities.item.WarpedItem;
 import warped.application.state.WarpedInventory;
 import warped.application.state.WarpedInventory.SortType;
-import warped.application.state.WarpedState;
+import warped.user.keyboard.WarpedKeyboard;
 import warped.user.mouse.WarpedMouse;
 import warped.user.mouse.WarpedMouseEvent;
 import warped.utilities.utils.Console;
@@ -43,6 +46,7 @@ public class GUIInventory<T extends ItemBindable<?>> extends WarpedGUI {
 	private Color hoverColor = new Color(60,60,60,60);
 	
 	private WarpedInventory<T> selectInvent;
+
 	
 	private static int dropIndex;
 	private static int dragIndex;
@@ -71,12 +75,12 @@ public class GUIInventory<T extends ItemBindable<?>> extends WarpedGUI {
 	public WarpedInventory<T> getInventory(){return selectInvent;}
 
 	/**Drop an item from the previous inventory into this inventory.
-	 * @param previousInventory - the inventory that the item is being taken from.
-	 * @param dragIndex - the index of the item in the previous inventory.
-	 * @param dropIndex - the index where the item is to be inserted into this inventory.
 	 * @author 5som3*/
-	public void dropItem(WarpedItem<?> dropItem) {
-		if(dropItem.getObjectID().getGroupID().getManagerID() != selectInvent.getGroupID().getManagerID()) {
+	public void dropItem() {
+		WarpedItem<?> dropItem = WarpedMouse.getDraggedItem();
+		if(dropItem.getObjectID() == null) {
+			Console.err("GUIInventory -> dropItem() -> the item has no ID -> ignore this error if moving half stack");
+		} else if(dropItem.getObjectID().getGroupID().getManagerID() != selectInvent.getGroupID().getManagerID()) {
 			Console.err("GUIInventory -> dropItem() -> dropItem is not from the same item set : " + dropItem.getItemType());
 			return;
 		}		
@@ -88,7 +92,7 @@ public class GUIInventory<T extends ItemBindable<?>> extends WarpedGUI {
 		
 		dragIndex  = -1;
 		dropIndex  = -1;
-		WarpedMouse.dropItem(true);
+		WarpedMouse.dropItem();
 	}
 	
 	/**Drop an item from the previous inventory into this inventory.
@@ -97,7 +101,9 @@ public class GUIInventory<T extends ItemBindable<?>> extends WarpedGUI {
 	 * @param dropIndex - the index where the item is to be inserted into this inventory.
 	 * @author 5som3*/
 	public void restoreItem(WarpedItem<?> dropItem) {
-		if(dropItem.getObjectID().getGroupID().getManagerID() != selectInvent.getGroupID().getManagerID()) {
+		if(dropItem.getObjectID() == null) {
+			Console.err("GUIInventory -> dropItem() -> the item has no ID -> ignore this error if moving half stack");
+		} else if(dropItem.getObjectID().getGroupID().getManagerID() != selectInvent.getGroupID().getManagerID()) {
 			Console.err("GUIInventory -> dropItem() -> dropItem is not from the same item set : " + dropItem.getItemType());
 			return;
 		}		
@@ -160,17 +166,24 @@ public class GUIInventory<T extends ItemBindable<?>> extends WarpedGUI {
 		pushGraphics();
 	}
 	
-	
+	boolean leftDragging = false;
 	@Override
-	protected void mouseEntered() {return;}
+	protected void mouseEntered() {
+		if(WarpedMouse.isLeftPressed() && WarpedMouse.isDraggingItem()) {
+			leftDragging = true;
+		}
+		
+	}
 
 	@Override
 	protected void mouseExited() {
+		leftDragging = false;
 		hoverX = hoverY = -1;
 		updateGraphics();
 		
 	}
 
+	@SuppressWarnings("unchecked")
 	@Override
 	protected void mouseMoved(WarpedMouseEvent mouseEvent) {
 		int hx = Math.floorDiv(mouseEvent.getPointRelativeToObject().x, itemSpacingX);
@@ -179,14 +192,29 @@ public class GUIInventory<T extends ItemBindable<?>> extends WarpedGUI {
 		if(hx < 0 || hy < 0 || hx >= columns || hy >= rows) return;
 		
 		int index = hx + hy * columns;
-		if(index >= selectInvent.getMemberCount()) return;
+		if(index >= selectInvent.getMemberCount()) index = selectInvent.getMemberCount() - 1;
 		
 		if(hoverX != hx || hoverY != hy) {
 			//invent.getMember(index).unhovered();
 			hoverX = hx;
 			hoverY = hy;
 			updateGraphics();
+		}	
+		if(WarpedMouse.isDraggingItem()) {			
+			if(WarpedMouse.isRightPressed()) {
+				if(WarpedMouse.getDraggedItem().getQuantity() > 1) {
+					selectInvent.addMember(new WarpedItem<T>((T)WarpedMouse.getDraggedItem().getItemType()));
+					WarpedMouse.getDraggedItem().consume();
+					updateGraphics();
+				} else dropItem();
+			}
+			if(leftDragging && !WarpedMouse.isPressed() && WarpedMouse.isDraggingItem()) {
+				dropIndex = index;
+				leftDragging = false;
+				dropItem();
+			}
 		}
+		
 	}
 
 	@Override
@@ -199,24 +227,54 @@ public class GUIInventory<T extends ItemBindable<?>> extends WarpedGUI {
 		int index = hx + hy * columns;
 		if(index >= selectInvent.getMemberCount()) return;
 		
+		WarpedItem<T> selectItem = selectInvent.getMember(index);
 		dragIndex = index;
-		WarpedMouse.dragItem(this, selectInvent.getMember(index));		
-		selectInvent.removeMember(index);
+		if(WarpedKeyboard.isKeyPressed(KeyEvent.VK_SHIFT) && selectItem.getQuantity() > 1) {
+			WarpedMouse.dragItem(this, new WarpedItem<T>(selectItem.getItemType(), selectItem.getQuantity() / 2));
+			selectItem.setQuantity(selectItem.getQuantity() - selectItem.getQuantity() / 2);
+		} else {			
+			WarpedMouse.dragItem(this, selectItem);		
+			selectInvent.removeMember(index);
+		}
 		updateGraphics();
 	}
 
+	@SuppressWarnings("unchecked")
 	@Override
-	protected void mousePressed(WarpedMouseEvent mouseEvent) {return;}
+	protected void mousePressed(WarpedMouseEvent mouseEvent) {
+		if(WarpedMouse.isDraggingItem() && mouseEvent.getMouseEvent().getButton() == MouseEvent.BUTTON3) {
+			if(WarpedMouse.getDraggedItem().getObjectID() == null) {
+				Console.err("GUIInventory -> mousePressed() -> item has no ID -> ignore if this is a half stack");
+			} else if(WarpedMouse.getDraggedItem().getObjectID().getGroupID().getManagerID() != selectInvent.getGroupID().getManagerID()) {
+				Console.err("GUIInventory -> dropItem() -> dropItem is not from the same item set : " + WarpedMouse.getDraggedItem().getItemType());
+				return;
+			}	
+			
+			int hx = Math.floorDiv(mouseEvent.getPointRelativeToObject().x, itemSpacingX);
+			int hy = Math.floorDiv(mouseEvent.getPointRelativeToObject().y, itemSpacingY);
+			
+			if(hx < 0 || hy < 0 || hx >= columns || hy >= rows) return;
+			int index = hx + hy * columns;	
+			dropIndex = index;
+			if(WarpedMouse.getDraggedItem().getQuantity() > 1) {
+				selectInvent.addMember(new WarpedItem<T>((T)WarpedMouse.getDraggedItem().getItemType()));
+				WarpedMouse.getDraggedItem().consume();
+				updateGraphics();
+			} else dropItem();
+		}
+	}
 
 	@Override
 	protected void mouseReleased(WarpedMouseEvent mouseEvent) {
+		if(mouseEvent.getMouseEvent().getButton() == MouseEvent.BUTTON3) return;
+		leftDragging = false;
 		int hx = Math.floorDiv(mouseEvent.getPointRelativeToObject().x, itemSpacingX);
 		int hy = Math.floorDiv(mouseEvent.getPointRelativeToObject().y, itemSpacingY);
 		
 		if(hx < 0 || hy < 0 || hx >= columns || hy >= rows) return;
 		int index = hx + hy * columns;
 
-		if(WarpedMouse.isDraggingItem()) {
+		if(WarpedMouse.isDraggingItem() && mouseEvent.getMouseEvent().getButton() != MouseEvent.BUTTON3) {
 			Console.ln("GUIInventory -> mouseReleased() -> dropping item at index : " + dropIndex);
 			if(index > selectInvent.getMemberCount()) {
 				Console.ln("GUIInventory -> dropIndex() -> drop index too high -> chaned to : " + dropIndex);
@@ -224,7 +282,7 @@ public class GUIInventory<T extends ItemBindable<?>> extends WarpedGUI {
 			}
 
 			dropIndex = index;
-			dropItem(WarpedMouse.getDraggedItem());			
+			dropItem();			
 			return;							
 		}
